resume_matcher/requirements.txt
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.0.0
openai>=1.40.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
